Add file to test out everything
This just tests that the replication file successfully runs and saves all of the necessary model output. It only takes 1,000 MCMC samples, and then makes 100 climate projections and calculates 100 SC-CH4 estimates for each of the scenarios. It does not calculate the BMA weights (which require larger samples or else they error out because of issues with the coavariance matrices).