Modify "run_climate_models" functions for CAR(1) likelihood
The CAR(1) likelihood has a different number of uncertain parameters compared to the combined i.i.d. and AR(1) likelihood approach for CO2 and CH4 concentrations. This adjusts the parameter indexing so the model functions are consistent with the new log posterior function in the create_log_posterior.jl file.